# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"InviteResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"created": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"project_id",
					"email",
					"created"
				],
				"type": "object",
				"additionalProperties": true
			},
			"InviteValues": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TemplateResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"rule": {
						"type": "string"
					},
					"template": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"environment_id": {
						"type": "string"
					},
					"created": {
						"type": "string"
					},
					"updated": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name",
					"rule",
					"template",
					"id",
					"project_id",
					"environment_id",
					"created",
					"updated"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TemplateValues": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"rule": {
						"type": "string"
					},
					"template": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"rule",
					"template"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TemplateSearchResults": {
				"properties": {
					"total_hits": {
						"type": "number",
						"format": "double"
					},
					"templates": {
						"items": {
							"$ref": "#/components/schemas/TemplateResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"total_hits",
					"templates"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EnvironmentResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"project_id",
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EnvironmentValues": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateDelReqReport": {
				"properties": {
					"id": {
						"type": "string"
					},
					"outstandingConfirmations": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"outstandingConfirmations"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateDelReqRequestBody": {
				"properties": {
					"resourceKind": {
						"type": "string"
					},
					"resourceId": {
						"type": "string"
					}
				},
				"required": [
					"resourceKind",
					"resourceId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GetDelReqReport": {
				"properties": {
					"expired": {
						"type": "boolean"
					},
					"backoffRemaining": {
						"type": "number",
						"format": "double"
					},
					"outstandingConfirmations": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"expired",
					"backoffRemaining",
					"outstandingConfirmations"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "retraced",
		"version": "1.9.1",
		"description": "A fully open source audit log service and embeddable UI easily deployed to your own Kubernetes cluster",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/admin/v1/project/{projectId}/invite": {
			"post": {
				"operationId": "CreateInvite",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InviteResponse"
								}
							}
						}
					}
				},
				"description": "Create an invite. Sends an invitation email to the user.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The invite resource with the invitee's email",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteValues",
								"description": "The invite resource with the invitee's email"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListInvites",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InviteResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all invites.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/invite/{inviteId}": {
			"delete": {
				"operationId": "DeleteInvite",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Delete an invite.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment id",
						"in": "path",
						"name": "inviteId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/templates": {
			"post": {
				"operationId": "CreateTemplate",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TemplateResponse"
								}
							}
						}
					}
				},
				"description": "Create a template. An overview of Template usage in Retraced can be found at\n\nhttps://boxyhq.com/docs/retraced/advanced/display-templates",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment id",
						"in": "query",
						"name": "environment_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The template resource to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TemplateValues",
								"description": "The template resource to create"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "SearchTemplates",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TemplateSearchResults"
								}
							}
						}
					}
				},
				"description": "Search templates. An overview of Template usage in Retraced can be found at\n\nhttps://boxyhq.com/docs/retraced/advanced/display-templates",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environmentId",
						"in": "query",
						"name": "environment_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The maximum number of results to return",
						"in": "query",
						"name": "length",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "How many results to skip",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/templates/{templateId}": {
			"delete": {
				"operationId": "DeleteTemplate",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"description": "Delete a template. An overview of Template usage in Retraced can be found at\n\nhttps://boxyhq.com/docs/retraced/advanced/display-templates",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The id of the template to delete",
						"in": "path",
						"name": "templateId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment",
						"in": "query",
						"name": "environment_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/environment": {
			"post": {
				"operationId": "CreateEnvironmentRequest",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EnvironmentResponse"
								}
							}
						}
					}
				},
				"description": "Create a new environment.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EnvironmentValues"
							}
						}
					}
				}
			}
		},
		"/admin/v1/project/{projectId}/environment/{environmentId}": {
			"delete": {
				"operationId": "DeleteEnvironment",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"description": "Delete an environment and all of its dependents.\nThis is only allowed if\n1) the environment is empty (i.e. it lacks any recorded events), or\n2) an outstanding \"deletion request\" has been approved.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment to be deleted",
						"in": "path",
						"name": "environmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/environment/{environmentId}/deletion_request": {
			"post": {
				"operationId": "CreateDeletionRequest",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateDelReqReport"
								}
							}
						}
					}
				},
				"description": "Create a resource deletion request and associated confirmation\nrequirements (as necessary).",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment",
						"in": "path",
						"name": "environmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDelReqRequestBody"
							}
						}
					}
				}
			}
		},
		"/admin/v1/project/{projectId}/environment/{environmentId}/deletion_request/{deletionRequestId}": {
			"get": {
				"operationId": "GetDeletionRequest",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetDelReqReport"
								}
							}
						}
					}
				},
				"description": "Get the current status of an outstanding deletion request.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment",
						"in": "path",
						"name": "environmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The id of the deletion request to look up",
						"in": "path",
						"name": "deletionRequestId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/environment/{environmentId}/deletion_confirmation/{code}": {
			"post": {
				"operationId": "ApproveDeletionConfirmation",
				"responses": {
					"200": {
						"description": "OK"
					}
				},
				"description": "Mark a deletion confirmation as received (i.e. approve it).",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment",
						"in": "path",
						"name": "environmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The confirmation code",
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/project/{projectId}/token": {
			"post": {
				"operationId": "CreateApiToken",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiTokenResponse"
								}
							}
						}
					}
				},
				"description": "Create a new API token.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The environment id",
						"in": "query",
						"name": "environment_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApiTokenValues"
							}
						}
					}
				}
			}
		},
		"/admin/v1/project/{projectId}/token/{apiToken}": {
			"put": {
				"operationId": "UpdateApiToken",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiTokenResponse"
								}
							}
						}
					}
				},
				"description": "Update an API token's fields.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The token to update",
						"in": "path",
						"name": "apiToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ApiTokenValues_"
							}
						}
					}
				}
			}
		},
		"/admin/v1/project/{projectId}/token/{tokenId}": {
			"delete": {
				"operationId": "DeleteApiToken",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"description": "Delete an API token.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT authentication",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The project id",
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The token to delete",
						"in": "path",
						"name": "tokenId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/v1/token": {
			"post": {
				"operationId": "CreateAdminToken",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminToken"
								}
							}
						}
					}
				},
				"description": "Create an admin token. An Admin Token can be used to impersonate a user,\nand will have access to all the user's projects, environments, and events.",
				"tags": [
					"Admin API"
				],
				"security": [],
				"parameters": [
					{
						"description": "Base64 encoded JWT",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/auditlog",
			"description": "Local"
		}
	]
}